third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-sizing/contain-intrinsic-size/auto-010-expected.txt

This is a testharness.js-based test.
FAIL Last remembered size supports multiple fragments assert_equals: Using last remembered size - fragments expected 1 but got 2
FAIL Last remembered size is updated when 2nd fragment changes size assert_equals: Using updated last remembered size - fragments expected 1 but got 2
Harness: the test ran to completion.
